Update README.md
This project aims to predict credit score delinquency using a machine learning approach. The motivation behind this project is the increasing importance of accurate credit risk assessment in financial institutions. By developing a predictive model, we can assist lenders in making informed decisions about lending to individuals or businesses.

**Problem Statement:** We want to predict whether a borrower is likely to default on their credit obligations, which is a binary classification task. To achieve this, we employed a dataset containing various features related to the borrower's financial and personal information.

**Approach:** We began by performing exploratory data analysis (EDA) to understand the characteristics of the dataset. Data cleaning, preprocessing, and feature engineering were then carried out to prepare the data for modeling. We evaluated multiple machine learning algorithms, including Logistic Regression, Random Forest, XGBoost, and Decision Trees, to identify the best-performing model.

**Results:** After thorough experimentation and evaluation, we found that the Random Forest algorithm outperformed other models in terms of accuracy, ROC-AUC, and F1 score. This model achieved an accuracy of 91.9%, suggesting its strong ability to predict credit score delinquency.

**Insights:** From the feature importance analysis, we discovered that variables such as "RevolvingUtilizationOfUnsecuredLines" and "Family_size" played a significant role in predicting credit delinquency.

By sharing this project, we hope to contribute to the field of credit risk assessment and provide a template for others looking to develop predictive models in the finance domain.

Feel free to explore the Jupyter Notebook and code files to dive deeper into the project's implementation and analysis.